+++ src/utils/consoleIO.ts
@@ -1,13 +1,19 @@
 import { selectDebugMode } from "./selectors";
 import type { Option } from "../game/interactions/option";
 
-export const output = (text: string) => {
-  console.log(text);
+export const output = (text: string, error?: boolean, warn?: boolean) => {
+  if (error) {
+    console.error(text);
+  } else if (warn) {
+    console.warn(text);
+  } else {
+    console.log(text);
+  }
 };
 
-export const debug = (text: string) => {
+export const debug = (text: string, error?: boolean, warn?: boolean) => {
   if (selectDebugMode()) {
-    output(text);
+    output(text, error, warn);
   }
 };